requirements.txt
bleach==1.4.3
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11.22
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.6.0
html5lib==0.9999999
idna==2.8
Markdown==2.6.6
Pillow==3.3.0
psycopg2==2.8.3
python-decouple==3.0
pytz==2019.1
requests==2.22.0
six==1.12.0
static3==0.7.0
Unipath==1.0
urllib3==1.25.3
whitenoise==3.2
numpy-1.16.4
pandas-0.24.2
python-dateutil-2.8.0
lightfm-1.15
scipy-1.2.2
scikit-learn-0.20.3
sklearn-0.0
recsys-0.0.4